feat: first working implementation

+++ pkg/util/configs/opts.go
@@ -10,11 +10,11 @@ var Opts struct {
 
 	Output string `short:"o" long:"output-file" description:"The name of the JSON file to persist the trained model to." default:"results.json"`
 
-	HiddenLayers int64 `short:"h" long:"hidden-layers" description:"The number of hidden layers to use. (default: 1)" default:"1"`
+	HiddenLayers int `short:"h" long:"hidden-layers" description:"The number of hidden layers to use. (default: 1)" default:"1"`
 
-	Nodes int64 `short:"n" long:"nodes" description:"The number of nodes per hidden layer. (default: 10)" default:"10"`
+	Nodes int `short:"n" long:"nodes" description:"The number of nodes per hidden layer. (default: 10)" default:"10"`
 
-	Rounds int64 `short:"r" long:"rounds" description:"The number of learning iterations (default: 1.000.000)" default:"1000000"`
+	Rounds int `short:"r" long:"rounds" description:"The number of learning iterations (default: 1.000.000)" default:"1000000"`
 
 	Speed float64 `short:"s" long:"speed" description:"The factor to be used to alter the weights of all nodes. (default: 0.1)" default:"0.1"`
 }
